requirements.txt
pyyaml==3.12
tornado
requests